+++ src/App.svelte
@@ -9,6 +9,7 @@
   import Drawer from './ui/components/Drawer.svelte';
   import LibraryPanel from './ui/components/LibraryPanel.svelte';
   import AudioEditor from './ui/components/AudioEditor.svelte';
+  import AssignBanner from './ui/components/AssignBanner.svelte';
 
   let { app }: { app: App } = $props();
 </script>
@@ -25,6 +26,10 @@
 
 {#if app.store.libraryOpen}
   <LibraryPanel {app} />
+{/if}
+
+{#if app.store.assigningSampleId}
+  <AssignBanner {app} />
 {/if}
 
 {#if app.store.audioEditor}